switch back to uint16 (lz)

+++ src/ERC721X.sol
@@ -10,11 +10,11 @@ contract ERC721X is ERC721, IERC721X, MinimalOwnable {
 
     address public minter;
     address public immutable originAddress;
-    uint32 public immutable originChainId;
+    uint16 public immutable originChainId;
     bytes4 constant interfaceID = 0xefd00bbc;
     mapping(uint256 => string) public _tokenURIs;
 
-    constructor(string memory _name, string memory _symbol, address _originAddress, uint32 _originChainId) ERC721(_name, _symbol) MinimalOwnable() {
+    constructor(string memory _name, string memory _symbol, address _originAddress, uint16 _originChainId) ERC721(_name, _symbol) MinimalOwnable() {
         minter = msg.sender;
         originAddress = _originAddress;
         originChainId = _originChainId;
